requirements.txt
Adafruit-Blinka==8.18.0
adafruit-circuitpython-busdevice==5.2.4
adafruit-circuitpython-requests==1.13.1
adafruit-circuitpython-typing==1.9.0
adafruit-circuitpython-vl53l0x==3.6.8
Adafruit-PlatformDetect==3.45.0
Adafruit-PureIO==1.1.10
autopep8==2.0.2
colorzero==2.0
future==0.18.3
gpiozero==1.6.2
iso8601==1.1.0
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==6.2.2
pycodestyle==2.10.0
pyftdi==0.54.0
pyserial==3.5
pyttsx3==2.90
pyusb==1.2.1
PyYAML==6.0
RPi.GPIO==0.7.1
serial==0.0.97
six==1.16.0
smbus2==0.4.2
sparkfun-qwiic-i2c==0.9.11
sparkfun-qwiic-scmd==0.9.1
sqlparse==0.4.3
typing-extensions==4.5.0
